feat : ajout de controlleur de transaction et fixer des erreures de payement

+++ SherElec-Back-end/DTOs/Request/TransactionRequest.cs
@@ -3,8 +3,8 @@
     public class TransactionRequest
     {
         public int OffreId { get; set; } // ID de l'offre
-        public int Amount { get; set; }  // Montant total en euros
-        public int Quantite { get; set; } // Quantité achetée
+        public double Amount { get; set; }  // Montant total en euros
+        public double Quantite { get; set; } // Quantité achetée
         public int AcheteurId { get; set; } // ID de l'acheteur
         public int VendeurId { get; set; } // ID du vendeur
     }
